git-lfs support for OSS repos (fixes #3851) (#3861)
The problem here is that when we slice out `tbn` into OSS repositories, we use it by applying diffs with `git am`, which lives under the git-lfs abstraction boundary, so we push pointers to files that don't exist in lfs-land for the public repositories.

After many false starts, I arrived at a solution I didn't hate and that worked at all: for each project we push, after we've pushed the code:

- add the public repo as a remote to tbn
- push all top-of-tree files to that remote using `git lfs push`
- remote the remote

This won't be 100% correct; if we haven't published for a while, there's a chance it won't push objects that were present in the changelist but no longer are. I think that's fine; each tagged release will be correct.

I did this in a shell script because was easy; the `osstool` go code just shells out to it, and I think that's fine.

Other changes here:

- add a note about git-lfs to the docs README
- make it so that we don't keep having to change the SHA for composer in the oss-push Dockerfile

fbshipit-source-id: 2bde7f0