new favicons for www, docs-site (#3925)
The favicon generator site has updated recommendations on the broadest possible compatibility with the fewest possible files. Those recommendations are implemented here.

We weren't actually serving a lot of these files correctly, because some of the files weren't imported at all and those that were got renamed based on their hashes. So for example, no root level `favicon.ico` was being served, because (1) it wasn't referenced in code anywhere, and (2) even if it had been it would have been named `<hex string>.ico` instead.

Here we change the file loader to special-case files with /favicon/ in their path, so that they are not renamed, and then we explicitly require them all in `entry.js`. With that in place, we can safely use vanilla meta and link tags, instead of needing to do inline requiring (which in any case rewrites the filename).

The `?v=<hexstring>` is so that when/if we rev the icons, we can change the hex string to force browsers to pull the new icon. I have seen this work some, but not all, of the time, but it seems harmless.

I also moved all the favicon files to `web-common` and fixed up docs-site to properly serve them as well.

Assuming we like this approach, and once we finialize which icons to use, I'll follow up with a similar change to move `app` on to `web-common/favicon`, so we only have to update this stuff once.

fbshipit-source-id: 0d842c3