[app] new houston logo, [app/docs/keycloak/www] consolidate logos and favicons (#3968)
First I was just going to add the new houston logo to the app. Then I got annoyed that we have logos all over the tree, and started to factor them out into `web-common`. On the way, I noticed that keycloak doesn't use our new favicon, but it does now. So the changes look like:

#### web-common
- patch `.eslintrc` to not blow up on long lines if they are long because of string literals
- add various logos
- fix `manifest.json` and `browserconfig.xml` to use the `?v=` hack

#### app
- remove old unused logos
- change ref to logo to `web-common` (this produced the long line that required the above .eslintrc change)

<img width="585" alt="screen shot 2017-10-04 at 2 59 45 pm" src="https://user-images.githubusercontent.com/362808/31201751-b41c7086-a914-11e7-84a5-67b5c63b5ee0.png">

#### docs

- remove old logo
- add logo from `web-common` in the `Dockerfile`
- remove unused (and confusing) `header.html` file

An unfortunately side-effect of this is that local builds no longer include the logo. There are a couple of approaches that fix this, but the `docs` project is open-sourced, and it's not cool to reference files outside of the project in a `Makefile`.

:cookie: `glen.2017-10-04.gds-app-new-houston-logo.cbfd02f1`

#### keycloak

- Remove all the old keycloak favicons, and the logo
- add favicons and logo from `web-common` in the `Dockerfile`
- use the `?v=` hack in all favicon references

Note that we don't do the mega-multi-format favicon dance here, because there's no way to place content at the root of the server. Nobody's bookmarking the login screen, stock favicon should be fine.

#### www

- remove the old logos
- replace imports of old logos with imports of new logos
- this somehow tickled a weird flow thing, which I could fix with `// $FlowIssue` but stack-overflowed my way to a fix, in which `.svg` are mapped to a non-existent module. This same trick did not seem to work in the app, which is why I left `// $FlowIssue` in place there.

fbshipit-source-id: 89a2e7d